+++ milvus/__init__.py
@@ -1,6 +1,8 @@
-from .client.Client import Milvus, Prepare, Status, IndexType
+from .client.GrpcClient import GrpcMilvus as Milvus
+from .client.GrpcClient import Prepare
+from .client.Abstract import IndexType
+from .client.Status import Status
 
 __all__ = ['Milvus', 'Prepare', 'Status', 'IndexType', '__version__']
 
-__version__ = '0.1.16'
-
+__version__ = '0.2.0'